remove now unused samplesheet components, rename column_component to header_component
