--- app/components/nextflow/samplesheet/dropdown_cell_component.html.erb
@@ -1,12 +1,0 @@
-<%= fields.select name,
-              values,
-              { required:, include_blank: true },
-              {
-                "aria-label": name,
-                class:
-                  "w-full focus:ring-primary-500 focus:border-primary-500 p-2.5 dark:bg-slate-700 bg-slate-50 border-none dark:border-none dark:placeholder-slate-400 text-inherit text-sm dark:focus:ring-primary-500 dark:focus:border-primary-500 cursor-pointer",
-                data: {
-                  "action" =>
-                    "change->nextflow--samplesheet#updateEditableSamplesheetData",
-                },
-              }.deep_merge(data) %>
