+++ app/components/nextflow/samplesheet/header_component.html.erb
