--- app/components/nextflow/samplesheet/metadata_cell_component.html.erb
@@ -1,19 +1,0 @@
-<% if metadata.present? %>
-  <div
-    class="p-2.5 text-nowrap"
-    data-controller="nextflow--autofilled-cell"
-    data-nextflow--autofilled-cell-input-name-value="workflow_execution[samples_workflow_executions_attributes][<%= index %>][samplesheet_params][<%= name %>]"
-    data-nextflow--autofilled-cell-input-value-value="<%= metadata %>"
-    data-nextflow--autofilled-cell-type-value="metadata"
-    data-nextflow--autofilled-cell-property-value="<%= name %>"
-  >
-    <span class="h-3.5 inline-block"><%= metadata %></span>
-  </div>
-<% else %>
-  <%= form.label name, class: "sr-only" %>
-  <%= render Nextflow::Samplesheet::TextCellComponent.new(
-    name,
-    fields: form,
-    required:,
-  ) %>
-<% end %>
