--- app/components/nextflow/samplesheet/metadata_cell_component.rb
@@ -1,25 +1,0 @@
-# frozen_string_literal: true
-
-module Nextflow
-  module Samplesheet
-    # Component to render a cell with sample metadata into the sample sheet
-    class MetadataCellComponent < ViewComponent::Base
-      attr_reader :form, :metadata, :name, :sample, :required, :index
-
-      # rubocop: disable Metrics/ParameterLists
-      def initialize(form:, name:, sample:, index:, field: nil, required: false)
-        @form = form
-        @name = name
-        @sample = sample
-        @required = required
-        @index = index
-        @metadata = metadata_value(field || name)
-      end
-      # rubocop: enable Metrics/ParameterLists
-
-      def metadata_value(name)
-        sample.metadata.fetch(name, '')
-      end
-    end
-  end
-end
