--- app/components/nextflow/samplesheet/text_cell_component.html.erb
@@ -1,9 +1,0 @@
-<%= fields.text_field(
-  name,
-  class:
-    "bg-slate-50 border-0 text-slate-900 text-sm focus:ring-primary-500  block w-full p-2.5 dark:bg-slate-700 dark:border-slate-600 dark:placeholder-slate-400 dark:text-white dark:focus:ring-primary-500 dark:focus:border-primary-500",
-  required: @required,
-  data: {
-    "action" => "change->nextflow--samplesheet#updateEditableSamplesheetData",
-  },
-) %>
