--- test/components/previews/nextflow_samplesheet_text_cell_component_preview/default.html.erb
@@ -1,4 +1,0 @@
-<%= form_with do |form| %>
-  <label class="sr-only" for="some[things]">Some things</label>
-  <%= render Nextflow::Samplesheet::TextCellComponent.new("some[things]", fields: form) %>
-<% end %>
